fix(bionic): variadic integers taken whole for the core to narrow; erfcf bound
* The adapter narrowed every signed variadic integer to 32 bits itself, so a
  `%ld` past 2^31 printed wrong; it now passes the whole slot and the core
  narrows by the length modifier. A real variadic `snprintf` call proves
  both halves: a `%u` with garbage above it is 256, a `%ld` of 5e9 is 5e9.
* erfcf, inline (283): M6, the renderer once its descriptor update
  templates existed, called it by name.
* The recvmmsg test gives msg_namelen room for a sockaddr_in6, so the 16 an
  IPv4 sender writes back is a value the call wrote and not the one it was
  handed.

+++ crates/omni-android/src/bionic/format.rs
@@ -136,9 +136,10 @@ fn collect(
     for (index, kind) in kinds.iter().enumerate() {
         let argument = first_variadic + index;
         owned.push(match kind {
-            // A variadic `int` occupies a whole 64-bit slot and the caller is not required to
-            // clear the high half, so it is taken as 64 bits and narrowed here.
-            ArgKind::Int => Owned::Int(source.next_u64()? as u32 as i32 as i64),
+            // A variadic integer occupies a whole 64-bit slot and is taken whole: **whether it is
+            // an `int` or a `long` is the length modifier's to say**, and the core narrows by it.
+            // Narrowing here printed every `%ld` of a value past 2^31 wrong.
+            ArgKind::Int => Owned::Int(source.next_u64()? as i64),
             ArgKind::UInt => Owned::UInt(source.next_u64()?),
             ArgKind::Ptr => Owned::Ptr(source.next_u64()?),
             ArgKind::Double => Owned::Double(source.next_f64()?),
